[audio_core] Add rounding mode to TimelineRate::Scale
In preparation for a fix to fxb/67996, add a rounding mode to Scale.
There are three options: Truncate, Floor, and Ceiling.

Prior to this CL, Scale always rounded using Floor. This is a bit
non-intuitive: with integers, we expect truncation. Floor and Truncate
differ when the input value is negative. I manually inspected every call
to TimelineRate::Scale. I found exactly one call where the input value
might be negative: the call within TimelineFunction::Apply. In all other
cases, the value is expected to be positive (usually this value is a
frame count or a positive duration). Hence, it is safe to use Truncate
in all of these calls. In TimelineFunction::Apply, we continue using
Floor for reasons specified in comments.

While I'm here, I also simplified TimelineRate's public API by removing
unused and redundant functions and methods.

This CL should be a no-op.

Bug: 67996
Change-Id: If22b406ac1d4b70713d2f1522e36e6bd3713ed4d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474703
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>