[sshutil][swd-e2e-tests] Introduce Resolver interface
SWD's e2e tests usually identify a device based off of a name derived from
the MAC address. Unfortunately this doesn't work well for link-local
devices for two reasons:

* Some link-local network configurations may generate a new interface name
  on a host-linux device every time a device is rebooted.
* If the link-local device address is IPv6, paving the device will change
  the IPv6 address.

Both of these can cause the OTA test to fail. The traditional way to handle
an issue like this is just to leverage DNS, but we can't use DNS with the
MAC-address based devices. So this introduces a notion of an abstract
"Resolver" which allows us to discover an address through DNS or the
MAC-based device name.

One thing to note is that this CL makes two resolvers, one for sshutil, and
one for swd-e2e-tests. That's because sshutil wants to resolve a concrete
address, but in the e2e tests, we want to resolve just the host name so
we could connect to potentially a few different ports on the device.

Change-Id: I1c158e1cb38d0317380e4cd2e57f3a1a595805da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/469862
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>