[audio_core] Fix overflow resulting from a rounding error
For justification for this change, see this comment:
https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=67996#c22

This CL adds a regression test which fails before this CL (by crashing
in PointSampler) and passes after this CL. The test I added relies on
making a direct call to MixStage::ProcessMix after directly setting some
values Mixer::source_info and Mixer::bookkeeping. This is ugly -- it
would be nicer to use the MixStage public API only -- however MixStage
has so much internal state that it's difficult to create a test case
which makes that internal state align in exactly the right way to
trigger the crash.

Bug: 67996
Change-Id: Icb6fad5ba3d3dc2a606fb5b0097d9b656fa54e35
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474704
Commit-Queue: Tom Bergan <tombergan@google.com>
Fuchsia-Auto-Submit: Tom Bergan <tombergan@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>