[bt][gap] Handle LECM inbound/outbound connection race
When a remote initiated connection is created in
LowEnergyConnectionManager during the connection procedure of a locally
initiated connection, return an error to the local connection requestor
instead of asserting.

Similarly, when a remote initiated connection is created while scanning
is in progress for a locally initiated connection and scanning times
out, return an error to the requestor instead of crashing.

Fixed: 68343
Test: in bt-host-unittests
GAP_LowEnergyConnectionManagerTest.RegisterRemoteInitiatedLinkDuringLocalInitiatedLinkConnecting
GAP_LowEnergyConnectionManagerTest.RegisterRemoteInitiatedLinkDuringLocalInitiatedConnectionScanning

Change-Id: I309bea991106eb84dce5be28bc0d83353002b78e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480139
Commit-Queue: Ben Lawson <benlawson@google.com>
Reviewed-by: Xo Wang <xow@google.com>