[audio_core] Better overflow treatment in Timeline{Rate,Function}
This CL makes the following changes:

1. Add TimelineRate::kUnderflow. This brings TimelineRate closer to
   affine::Ratio, which is another small step towards our long-term goal
   of replacing TimelineRate with affine::Ratio.

2. Add sanity checking to AudioRenderer::Play to ensure that
   TimelineRate and TimelineFunction operations cannot overflow.

Broadly speaking, we use TimelineRates and TimelineFunctions for the
following kinds of transformations:

  {frames, bytes, time} -> {frames, bytes, time}

Throughout audio_core, we have a widespread assumption that TimelineRate
and TimelineFunction operations do not overflow because the inputs are
"realistic" values. For example, given that:

1. A frame timeline starts at zero and advances by at most 192k fps
   (which is 192k * 2^13 = 1.572e9 fractional frames per second).

2. The monotonic clock timeline starts at boot and advances one tick per
   nanosecond.

We can translate from frame numbers to timestamps using the function
start_time_seconds + frame_number * (192k * 2^13). We have at least 185
years of up time before we might possibly overflow this function.

A principled way to address this problem is to define a maximum up time
for each AudioRenderer (e.g., one year), then each time we create a
TimelineFunction or TimelineRate, we would simulate advancing that
function (and its inverse) the equivalent of one year in the future and
validate that the function remains computable without overflows.

Doing this consistently throughout audio_core would require quite a
large refactor. In lieu of that, for now, we perform sanity checks in
the following places to ensure that TimelineFunction inputs are
"realistic":

1. When setting up the driver ring buffer. A sanity check already exists
   here:
https://cs.opensource.google/fuchsia/fuchsia/+/master:src/media/audio/audio_core/audio_driver_v2.cc;drc=ec1d12b3b41f449eeda4dd1140218a33f2559fbc;l=334

2. When receiving a frame rate in SetPcmStreamType. A sanity check
   already exists here:
https://cs.opensource.google/fuchsia/fuchsia/+/master:src/media/audio/lib/format/format.cc;drc=78e3d6b1504efb6b4b44ddd83e0625b3660b506b;l=28

3. When receiving a payload buffer. We assume the kernel has
   sanity-checked the payload VMO to be something that can fit in memory
   without OOMing, which limits VMO size to about ~2^33 bytes on typical
   consumer devices (in practice much less).

4. When receiving a timestamp in AudioRenderer::Play().

This CL adds sanity checks in the last case.

These sanity checks do not guarantee that TimelineFunctions will not
overflow within a MaxUpTime, however they should be good enough to avoid
overflow in practice.

Change-Id: I283e1e0322c0f3ce151586cae1fe4b6995ac2208
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477058
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>
API-Review: Dale Sather <dalesat@google.com>