[perfcompare] Specify precision explicitly for printing point estimates
This preserves the behaviour that Python 2 gives and hence keeps the
golden file tests passing when they are run on Python 3 (with other
fixes applied).

On Python 2.7, str() on a float is equivalent to formatting with
'%.12g' while repr() on a float is equivalent to formatting with
'%.17g'.

By Python 3.8, that has changed so that str() and repr() on a float
both use '%.17g'.

Python's behaviour was changed in this issue:
https://bugs.python.org/issue9337

The older behaviour was documented here:
  "repr(float) produces 17 significant digits"
  "Python’s builtin str() function produces only 12 significant digits"
https://docs.python.org/3.0/tutorial/floatingpoint.html

Bug: none
Test: ./prebuilt/third_party/vpython/vpython garnet/bin/perfcompare/perfcompare_test.py
Change-Id: I07050dc56e759b5759b8292b5fe017799ae4db96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481023
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>