[kernel][asan][arm64] Build / boot support
Basic support for building {bringup...}.arm64 --variant kasan;
the system boots, allocates the kernel ASAN shadow 'late', and has
out-of-line asan checks, but does not mark ASAN as initialized yet.

We move certain parts of kernel asan initialization from generic code
to architecture-specific code; in the x86 implementation of kASAN, we
construct page tables for the kASAN shadow manually in the loader
assembly. In the arm64 implementation, we use standard VM interfaces
instead - this means we cannot enable kernel ASAN until later in the
system lifecycle.

In order to support 'late' shadow initialization, we disable inline
instrumentation on arm64; the outline instrumentation can check an
initialized flag. When we move to allocate the shadow early, we can
re-enable it.

Tested:
* 600 runs of 'k ut pi' && 'k ut thread' && 'k ut timer' under qemu,
  bringup.arm64 --variant kasan
* 600 runs of 'k ut pi' && 'k ut thread' && 'k ut timer' under kvm,
  bringup.x64 --variant kasan

Bug: 30033 Kernelmode ASAN for Zircon
Pair-programmed-with: mvanotti@

Change-Id: Id7659942e5d3ad4fdd1b59dc1cf3dbcc2221f888
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388863
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>