[build][tracer] Failures should not freshen output
Check that failed actions do not leave an output that appears
to be up-to-date relative to inputs.

Note that depfiles at this time are still allowed but not required,
which means that writes to depfiles from failed actions will not be
caught with the current logic.

For reference, automake does this safely by writing a depfile to a
temporary file and then moving it in-place after the compile command
succeeds.  This way the depfile always reflects the last good compile.

This is not expected to have any impact on healthy/green builders,
but may alert users with locally failed actions that something is wrong.

Test: fx test build/tracer

Change-Id: I7a231dbd2cb4c16876367a3087e85b28f9a1dd7b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477605
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>