[stdcompat] add cpp20::bit_cast.
The bit_cast polyfill is a 3-way branch, were std::* is picked first,
if this is not available we use bit_cast builtin. As a last resort we
introduce a non constexpr bit_cast.

The usage of aligned_storage, for memcpy and then reinterpret_cast,
allows us not to introduce a default_constructible requirement on type
To.

Introduced a LIB_STDCOMPAT_NONCONSTEXPR_BIT_CAST, defined only when the
last resot branch is picked. This is mostly for testing constexpr
scenario when this is not defined.

Bug: 57356
Test: stdcompat-tests
Change-Id: I1cdfa9bb97330391d45e9fd9a71d96621bf3fed0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478620
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>