[fshost] Proxy BlobfsVerifier FIDL from blobfs
Exposes the fuchsia.update.verify.BlobfsVerifier FIDL interface out of
fshost, which is implemented as a proxy service node.

Refactored fs-manager to not take ownership of the root export channel
in ForwardFsDiagnosticsDir, so that we can store the root export channel
in FsManager and reuse it for as many service proxies as we'd like.

Bug: 64608
Tested: Integration test
Tested: Manually verified the service is exported: `cs info -f fshost`
Change-Id: I279fd4088e21378cabcb4fe9062b3be083722b98
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481625
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>
Reviewed-by: Chris Suter <csuter@google.com>