[SetUI] Use Rebroadcast in Policy.
Currently the policy handler generates its own
settings change event in order to indicate when the
value has changed due to changes in policy. This is
coupled tightly with the switchboard logic, which is
being deprecated.

This change moves the notification responsibility to
the setting proxy via the rebroadcast command. The
setting proxy is then able to use its standard
communication method to convey the change.

The service MessageHub has been plumbed to the
policy handler as part of this change so that the
policy handler can invoke the rebroadcast command
on the proper MessageHub.

Test: fx test -o setui_service_tests
Bug: 68489
Change-Id: I10078ac21d8c9a4056c51beabf0f3066428eaee9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481011
Commit-Queue: Bryce Lee <brycelee@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: William Xiao <wxyz@google.com>