[go] Commit go.sum
It's essentially noise given our configuration but it's required by go1.16.
To clarify, we map Go dependencies via jiri (many are forked) and use
replace statements in go.mod.

I confirmed the file to be up to date with:
  for i in */go.mod; do G=$(dirname $i); echo $G; cd ./$G; go mod tidy; cd -; done

and confirmed resolving the go1.16 induced error with:
  cd tools
  # Edit go.sum and remove all references to cloud.google.com/go/storage
  go mod download cloud.google.com/go/storage

and the file became as before.

For more information see
- https://github.com/golang/go/issues/41934
- https://github.com/golang/go/issues/40728

The TL;DR: is that -mod=readonly becomes the default and when it's set,
go.sum must be present and clean.

This change has no effect on the GN build.

Bug: 69326
Change-Id: I50684c14564d7ac370eaf890776d233d18b5d5d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482481
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Commit-Queue: Marc-Antoine Ruel <maruel@google.com>