[jq5] Write basic jq5 tool.
Implements a minimal version of jq5, extending jq functionality to json5
objects. Future versions will aim to preserve comments, but this
implementation strips them. Files in json5 format are read by the
`read_json5` function and a tuple is returned  with a "json-ified" string
of the object along with a ParsedDocument representation including
comments. The `run_jq` function takes the strings and runs `jq` on them
with a provided filter.

Change-Id: I6392ae09724c00fdc463f2a93cc901a65ee64efc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476938
Commit-Queue: David Tamas-Parris <davidatp@google.com>
Reviewed-by: David Tamas-Parris <davidatp@google.com>
Reviewed-by: Shai Barack <shayba@google.com>