[kernel][phys] Refactor memory allocation
Use a bespoke owning type rather than unique_ptr specialization.
The custom deleter object storing the size means the UniquePtr
object is already two words and knows the size, but not in a
convenient fashion.  This made a separate object holding the size
be three words instead of two, with the same size stored twice.
The bespoke object just holds it as a span and makes it easy to
access so no separate object is needed, while also handling the
ownership semantics.

The Allocation object is now the only entry to the allocator API.
It uses fbl::AllocChecker to enforce error checking.

This separates the general allocation code from the code specific
to reading the ZBI, so it can be used in different phys contexts.

Bug: 68762
Change-Id: Iaf205c08161cccfc676b5f0c4b7517554d393b33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478857
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>