[kernel][phys] Decompressing ZBI chain-loader
This adds tests of decompressing a whole ZBI embedded in a ZBI item
and booting that ZBI.  So far it only works for arm64, where no page
tables are needed.  This is a more strenuous test of decompression
as physboot will use it, and it's also a test of arch::ZbiBoot that
boot shims (and interim physboot) will use.

Bug: 57764, 68762
Change-Id: I7108fb93480b0eebcfd0d1821083cb045606cf08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478720
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>