[media][audio_core] Update fuzzer deps
The audio-fuzzer/api-fuzzer is crashing due to missing deps and services
that bring in the fuchsia.media.AudioTestDevmgr service.

Locally, these failures are masked due to the inclusion of the services
and deps in encompassing packages; because fuzzer packages are extracted
as isolated units, these services and deps need to be explicitly
included.

Bug: 65547

Test: fx fuzz audio-fuzzer/api-fuzzer
Change-Id: I832327368dfe38b60448dbb4ceb52623a86540b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482657
Commit-Queue: Payton Bradsky <pbradsky@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>