[nelson][selina] Add selina device in its own devhost
This prevents an issue where the selina driver could become stuck due to
it being in the same devhost as the GPIO driver. SPI uses GPIO as well,
so there is a circular dependency if the selina and GPIO drivers are in
the same devhost.

Bug: 67567
Test: Radar CLI no longer hangs waiting for frames
Change-Id: I9eb108a71fa6e74a509c8a456280dba00cd2b30f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482497
Commit-Queue: Braden Kell <bradenkell@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>