[debugger] Syntax highlight source code listings.
Uses the tokenizer to provide simple syntax highlighting in source code
listings. This helps improve code readability significantly.

The syntax highlighter is limited to what the current tokenizer
supports, so there are no comments or C character literals, but this
will improve as we enhance the tokenizer.

A list of keywords for Rust and C is added since the tokenizer doesn't
split apart each keyword as a separate token.

Fixes a bug in the tokenizer where an underscore was treated as a
separator if immediately following a keyword. A test is added for this.

Change-Id: I6dd29557e338b854345384514ac12c6e48fc90c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482682
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>