[fidl] Generate dangerous ident tests at build time
This removes the checked-in generated dangerous identifier tests. All of
the tests are now generated at build-time.

Now that tests are checked in we can drop the 'tag' from identifier
definitions - they were just there to minimize diffs for the checked in
tests. Without the tags it's possible to shard the tests so that each
individual generated type is smaller - so this shards them into four.
With sharding it's possible to re-enable some tests that were previously
disabled because of build performance.

The deny listing is now almost entirely done in identifiers.py, but
there are some issues with Rust bindings deny listing and structs so the
struct tests are explicitly disabled for Rust in BUILD.gn

Change-Id: Id55b74175dd7b4fa6473169a1e10c7cc196d99bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443697
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Testability-Review: Benjamin Prosnitz <bprosnitz@google.com>