[kernel][vm] Support discarding pages
Implement discard operation that frees up all the pages of a reclaimable
discardable VMO, and sets the state to "discarded". Future unlocked
accesses on the discarded pages will fail. This is implemented by
checking the current discardable state in GetPageLocked(), and failing
the call if the state is discarded. VMOs will transition from the
discarded state to the unreclaimable state on the next lock operation.

This commit just introduces the ability to discard a VMO, but that
functionality is not used yet outside of tests.

Test: kernel-unittests

Bug: 60237
Change-Id: I8d8845068e388347b0cd8aecc0e30fb6cb354582
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476158
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>