[kernel][vm] Query discardable page counts
Query locked and unlocked page counts over all discardable VMOs in the
system. This requires iterating over the global discardable lists of
VmCowPages, which requires the DiscardableVmosLock; and querying the
page counts on each object, which requires the VmCowPages lock. Since we
cannot acquire the VmCowPages lock while holding the DiscardableVmosLock
(to avoid lock ordering deadlocks), we use the same strategy as for
traversing the all_vmos_ list.

The VmoCursor class is templated so it can be used both for all_vmos_
and the discardable lists. The DiscardableVmosLock is dropped between
iterations of the discardable lists, so that each individual VmCowPages
can be queried outside of the DiscardableVmosLock.

The discardable page counts can be printed with 'k scanner dump'. They
will also be exposed later through kmem stats.

Test: kernel-unittests

Bug: 60237
Change-Id: If9a6f45131d81437e4887e4deaf1daa8b9b721cb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478497
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>