[lapis] Rename "EventCount" to "Occurrence".
EventCount in Cobalt v1.0 is an overloaded type that
acts as a bucket for the number of times a thing happens, and
for raw integers. In v1.1 this ambiguity is addressed by introducing
explicit "Integer" and "Occurence" types. This naming is much clearer
and Lapis should use it to make the V1.1 cobalt migration more apparent.

Bug: 42067
Change-Id: I2d5eeccf3563978d40e03836edd6ad4c813a2098
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483240
Commit-Queue: Luke Nicholson <lukenicholson@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>