[debugger] Fix some TLS printing problems.
TLS access was broken when ELF symbols were required to be explicitly
annotated as such to be found by symbol lookup. This annotates the TLS
symbols as being ELF symbols.

Fixes an asan use-after-free error around the DwarfExprEval that is used
to execute the TLS program. This is unfortunately delicate that the
DWARF TLS thing can't conveniently use the simpler ASyncDwarfExprEval
class that would avoid this problem.

This TLS integration can't easily be tested without an end-to-end
integration test which is not currently practical.

TLS is still broken on x64 due to the fs.base register being missing.

Bug: 53940
Change-Id: Ie5ac8675bf9806a23e3ad10e5c6ce57d69110601
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483017
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>