[journal] Change the fuzzer to avoid hang
The journal's current API means that if you call WriteData and there's
not enough room in the buffer to store the operations, it will block
until there is room, but if the caller hasn't done anything with the
previous reservation, it will just deadlock. To fix this properly, we'd
need to substantially modify the journal API so that it keeps track of
which reservations have been submitted and which haven't been, but
that's tricky.

For now, just fix the fuzzer so that it doesn't send two data
operations, only one, and that should never block.

Fixed: 66376
Change-Id: I3ce3ed2cb65ddbddd4452dbeeee33b23ad1eba24
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483011
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Commit-Queue: Chris Suter <csuter@google.com>