[scrutiny] Add configurable build path.
This change adds a configurable build path option to Scrutiny. This
allows the build path to be fixed instead of retrieved the environment
which is useful for build scripts or custom builds. Setting the
build_path will let you define where the model collects data from.

This change also integrates a new testing library into Scrutiny which
removes a lot of duplication of the testing code and was required to
simplify testing the new ModelEnvironment integration. In addition this
change cleans up package data collection to improve its readability
which was one of the core consumers of the build path from the
environment.

Bug: 68938
Change-Id: Ibd2fbd333e814066baf0ff4fa22f520fd1c291a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481016
Reviewed-by: Drew Fisher <zarvox@google.com>
Reviewed-by: Yu Shan <shanyu@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>