[rustc_binary] make output_name support "-"
If output_name is passed to rustc_binary, that becomes the output_name
of the executable, without "-" -> "_" normalization, as one might expect.
The crate_name of the binary (observable with `module_path!()` )
continues to be normalized.

A future CL will apply this same logic to the target name itself, but
output_name is part of the transition path.

Bug: 64100
Change-Id: I02369de9b6162555139616b97a9a1d9b72da762e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450757
Commit-Queue: Dan Johnson <computerdruid@google.com>
Reviewed-by: Shai Barack <shayba@google.com>