[scrutiny] Add an in-memory store to Scrutiny.
This small change allows consumers to specify a DataModel uri of
"{memory}" which will utilize the MemoryStore instead of the
EmbeddedStore. With this change "{memory}" is also set as the default
model uri. The reason for this change is beacuse for most workloads
Scrutiny is used in the results are wanted always on the latest build of
data instead of on an existing model of the data. Housing this data in
memory for the lifetime of script is only several megabytes in the
largest case for current workloads.

If a custom workload does require a specific model path to be built
consumers already have access to the -m command to specify a custom path
or if they are code consumers they are able to do this through the
frontend configuration by simply specify a path.

Change-Id: If486a876667f6c496ff616f271205241f1d1b651
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482197
Reviewed-by: Drew Fisher <zarvox@google.com>
Reviewed-by: Yu Shan <shanyu@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>