[scrutiny] Add command engine.model.environment
This small change adds a controller that exposes the model environment.
This is useful for debugging consumer issues when they aren't getting
their expected results from their build as we can verify through a
command that the model environment is setup correctly. Such as the build
directory is at the correct path and the data models uri is correct.

Change-Id: Idcaaa861fba7a225d68c76b4c74590c1e0178b6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482198
Reviewed-by: Drew Fisher <zarvox@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>