[scrutiny] Add meta to core::Package collection definitions.
Currently packages only support contents which isn't all of the data
available in a package. Inline metadata can also be included directly
into the FAR. This data should be captured by the Package collection as it is
important for analysis. For example quickly examining all the allowlists
in the meta/ of a package.

Change-Id: I0eb544ca1a8ae4049d2e514807e671ee16634403
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482199
Reviewed-by: Drew Fisher <zarvox@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>