[codec] Allow double padding.
The logic to decide if it's ok to accept more input
was from back when we had a larger stream buffer.

Fix the logic to allow adding more input to the
stream buffer if the specific input fits instead of
assuming worst-case potential input size.  If the
actual input doesn't fit in the stream buffer this
time, we keep it around until it does fit.  This
way we avoid penalizing ourselves for having some
extra headers / in-band stream padding in the stream
buffer already, which could potentially double-count
already-"parsed" (in "parser" HW sense) padding + max
potential padding in a packet.

It's worth pointing out that if a stream has an
arbitrarily large amount of padding between frames
and is also protected preventing us from skipping the
padding in SW, there's no way for us to size the stream
buffer large enough to decode that on this HW.  In
practice it shouldn't be a problem though.

Also use 2 * kPaddingSize when sizing the stream
buffer, since not all OOB padding we add after a first
frame is necessarily consumed before parsing a
second frame and its OOB padding.

BUG: b/175801498
Change-Id: I3c19d5d8c495bac1bac7d84d4a52d23947d81230
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482817
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>