[libc] Build workaround to unstick gwp_asan roller
The gwp_asan sources removed a file, so the next integration roll
can't land until the build is changed.  But the build can't be
changed to omit the old file until the inregration roll lands.
So split the difference with a kludge to detect the old pin state
and adapt the build to both old and new.  The conditional hack
can be removed as soon as the integration roll lands.

Change-Id: Iaa6ea9ea471504d3c3da2ba6865a508795b9e2d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483349
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Kostya Kortchinsky <kostyak@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>