[stdcompat]: Add to_address polyfill.
std::span spec relies on to_address for multiple places, so I'm using
adding a polyfill for that. If its available we'll use that.

Noticed that we were not using the polyfills in on test, so it builds
even when the std version is not available.

Test: stdcompat-tests
Change-Id: Ibfa90f52c7936259b5802c186719a6c4894fe462
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482899
Reviewed-by: Roland McGrath <mcgrathr@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>