[pager] Handle VMO freeing from the VFS pager
Note: The VFS pager is not currently used. This patch fixes cleanup of
VMOs when no clones exist.

This patch notices when there are no more memory mappings of the VMO and
frees the associated object in the Vnode. Registration of paged vnodes
is changed to only track vnodes with mapped pages.

There are still some issues around the PagedVnode::vfs_ backpointer that
need resolving.

Change-Id: I1c86d5c6e6408ecf453053f262b4913ce471778c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478359
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>