[archivist] Switch to service config files instead of args.
We want to support different sets of diagnostics analysis services on
device depending on system build parameters. This change makes the
Archivist look in a specified location for a new services_config.json
file, which contains a list of services to connect to.

We connect to these services when Archivist starts up to support "lazy
binding." With this change, we will be able to create product
configurations that exclude certain services.

Bug: 69296
Change-Id: I9f07fec2ee6e014b80d1308f3105533aa2f9ceb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483460
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Shai Barack <shayba@google.com>