[cm][resolvers] Send fuchsia.mem.Data instead of ComponentDecl in Resolver protocol
Component manifests have lots of repeated string data, and can grow to
be larger than a zircon channe messagel size. Therefore, it is not a good
idea to send the ComponentDecl FIDL directly in a FIDL response.

Component resolvers typically deserialize the ComponentDecl from the
component's package, but don't actually read it. So it seems just forwarding
the uninterpreted bytes is the best solution. To account for large
manifest sizes, we use fuchsia.mem.Data, which is a union that can
either contain the bytes inline, or a handle to a VMO.

For base-resolver and universe-resolver, this works well, since the VMO
can be extracted straight from the package and forwarded to component
manager, not needing to read the contents in the resolver.

For future resolvers whose components don't have packages, and whose
manifests are small, the resolver can send the manifest inline.

Fixed: 69312
Change-Id: Iee3228d06f6274712e6895cc78aaf0ba1f5b0d6b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482757
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Adam Lesinski <adamlesinski@google.com>
API-Review: Gary Bressler <geb@google.com>